requirements.txt
python-dotenv
jira
python-telegram-bot
httpx